more execution time tests
