+++ api/dbutils/timer.go
@@ -1,0 +1,29 @@
+package dbutils
+
+import (
+	"log"
+	"runtime"
+	"time"
+)
+
+func callerName(skip int) string {
+	const unknown = "unknown"
+	pcs := make([]uintptr, 1)
+	n := runtime.Callers(skip+2, pcs)
+	if n < 1 {
+		return unknown
+	}
+	frame, _ := runtime.CallersFrames(pcs).Next()
+	if frame.Function == "" {
+		return unknown
+	}
+	return frame.Function
+}
+
+func Timer() func() {
+	name := callerName(1)
+	start := time.Now()
+	return func() {
+		log.Printf("%s took %v\n", name, time.Since(start))
+	}
+}
